+++ firmware/src/include/device.h
@@ -44,17 +44,17 @@ class SPI
         SPIClass* spi();
 
         void set_spi(SPIClass* spi);
-        void transfer(int channel, uint8_t data);
-        void transfer(int channel, uint8_t data[], int size);
+        void transfer(uint8_t channel, uint8_t data);
+        void transfer(uint8_t channel, uint8_t data[], uint16_t size);
         void commit(bool debug_out, SPIData* result);
 
     private:
         SPIClass* p_spi;
         SPIData* p_data;
-        int m_transfer, m_channel;
+        uint8_t m_transfer, m_channel;
 
-        void _on(int channel);
-        void _off(int channel);
+        void _on(uint8_t channel);
+        void _off(uint8_t channel);
 };
 
 #endif // DEVICE_H_INCLUDED
